requirements.txt
blinker==1.4
cryptography==3.4.8
pyjwt==2.1.0